Make lifePercent be quint8, so it always takes one byte in the QDataStream

+++ src/PokemonInfo/battlestructs.h
@@ -52,7 +52,7 @@ public:
     void removeStatus(int status);
 
     void init(const PokeBattle &poke);
-    virtual int lifePercent() const { return m_prop_lifePercent; }
+    virtual quint8 lifePercent() const { return m_prop_lifePercent; }
     quint8 &lifePercent() { return m_prop_lifePercent; }
 private:
     quint8 m_prop_lifePercent;
@@ -92,7 +92,7 @@ public:
 
     bool ko() const {return lifePoints() == 0 || num() == Pokemon::NoPoke || status() == Pokemon::Koed;}
     bool isFull() const { return lifePoints() == totalLifePoints(); }
-    int lifePercent() const { return lifePoints() == 0 ? 0 : std::max(1, lifePoints()*100/totalLifePoints());}
+    quint8 lifePercent() const { return lifePoints() == 0 ? 0 : std::max(1, lifePoints()*100/totalLifePoints());}
 
     void setNormalStat(int, quint16);
 private:
